+++ bot/loaders/__init__.py
@@ -1,5 +1,6 @@
 from .pdf import load_pdf
 from .ptt import load_ptt
 from .singlefile_html import load_singlefile_html
+from .url import load_url
 from .video_transcript import load_video_transcript
 from .youtube_transcript import load_youtube_transcript
