+++ tests/test_utils.py
@@ -3,24 +3,6 @@ from langchain_core.messages import AIMessage
 
 from bot.utils import ai_message_repr
 from bot.utils import parse_url
-from bot.utils import replace_domain
-
-
-@pytest.mark.parametrize(
-    "url, expected",
-    [
-        ("https://twitter.com/someuser/status/1234567890", "https://api.fxtwitter.com/someuser/status/1234567890"),
-        ("https://x.com/someuser/status/1234567890", "https://api.fxtwitter.com/someuser/status/1234567890"),
-        ("https://example.com/someuser/status/1234567890", "https://example.com/someuser/status/1234567890"),
-        ("twitter.com/someuser/status/1234567890", "twitter.com/someuser/status/1234567890"),
-        (
-            "https://subdomain.twitter.com/someuser/status/1234567890",
-            "https://subdomain.twitter.com/someuser/status/1234567890",
-        ),
-    ],
-)
-def test_fix_twitter(url, expected):
-    assert replace_domain(url) == expected
 
 
 @pytest.mark.parametrize(
